fix: moved github stats to different folder
